+++ backend/zane_api/views/serializers.py
@@ -362,6 +362,24 @@ class ProjectUpdateRequestSerializer(serializers.Serializer):
             )
         return attrs
 
+# ==============================
+#       Projects Search        #
+# ==============================
+
+class ProjectSearchSerializer(Serializer):
+    id = serializers.CharField(required=True)
+    created_at = serializers.DateTimeField(required=True)
+    slug = serializers.SlugField(required=True)
+    type = serializers.ChoiceField(choices=["project"], default="project")
+
+
+class ServiceSearchSerializer(Serializer):
+    id = serializers.CharField(required=True)
+    project_slug = serializers.SlugField(required=True)
+    slug = serializers.SlugField(required=True)
+    created_at = serializers.DateTimeField(required=True)
+    type = serializers.ChoiceField(choices=["service"], default="service")
+
 
 # ==============================
 #       service Update         #
